changes circleci to upload a single packed workflow to a release
Adds PyGithub to the requirements. This is used to manage github
releases. There are two methods a user can use to have circleci
pack a versioned workflow.
- Add a tag ending in '-v<number' to a commit and push to github
- Create a github release with tag version ending in '-v<number'

The part of the tag before the '-v<number>' part is assumed to
be the name of a workflow within the `workflows` directory.
This workflow will be packed and added to the release.
If there release doesn't exist the release will be created.